+++ src/main/index.ts
@@ -62,6 +62,7 @@ import { registerAwsResourceIpc } from '../core/aws-resource-register-ipc'
 import { registerTorrentIpc } from '../core/torrent/register-ipc'
 import { registerVirtualMachineIpc } from '../core/virtual-machine/register-ipc'
 import { registerCalendarIpc } from '../core/calendar/register-ipc'
+import { registerCdkHandlers } from './aws/cdk-manager'
 import { registerCloudflareCoreManagersIpc } from '../core/cloudflare-core-managers'
 import { registerHomeAssistantIpc } from '../core/home-assistant/register-ipc'
 import { registerHomeAssistantControlIpc } from '../core/home-assistant-control/register-ipc'
@@ -2391,6 +2392,7 @@ app.whenReady().then(async () => {
   registerTorrentIpc(corePlatform)
   virtualMachineManager = registerVirtualMachineIpc(corePlatform).manager
   registerCalendarIpc(corePlatform)
+  registerCdkHandlers(ipcMain, process.resourcesPath, app.getPath('userData'))
   registerCloudflareCoreManagersIpc(corePlatform)
   registerHomeAssistantIpc(corePlatform)
   registerHomeAssistantControlIpc(corePlatform)